Add signing rewritten assemblies on Full Framework
This commit adds the ability to sign the rewritten assembly
when running on Full Framework. Signing is not currently
supported by Mono.Cecil on Net Core.

When a keyfile is supplied and merge is not specified (false),
the keyfile will be used to sign the rewritten assembly. When
merge is specified, the keyfile will be used to sign the merged
assembly.

Add net472 as a TargetFramework to allow signing functionality.

Move from Mono.Options to CommandLineParser as AssemblyRewriter
is signed and Mono.Options is not signed, resulting in compilation
failing when targeting net472.