Add signing rewritten assemblies on Full Framework (#7)
* Add signing rewritten assemblies on Full Framework

This commit adds the ability to sign the rewritten assembly
when running on Full Framework. Signing is not currently
supported by Mono.Cecil on Net Core.

When a keyfile is supplied and merge is not specified (false),
the keyfile will be used to sign the rewritten assembly. When
merge is specified, the keyfile will be used to sign the merged
assembly.

Add net472 as a TargetFramework to allow signing functionality.

Move from Mono.Options to CommandLineParser as AssemblyRewriter
is signed and Mono.Options is not signed, resulting in compilation
failing when targeting net472.

* Update Mono.Cecil to support signing on Net core

This commit updates Mono.Cecil to support signing on .NET Core

* Allow multiple instances of command line flags

* Rename explicit interface properties and methods

* Additional Rewrites

This commit adds additional rewrites for

1. explicitly implemented interface properties
2. explicitly implemented interface methods

Refactor the common name replacement. This can
probably be refactored further.

Verified rewriting with ildasm to il and searching for old names
Verified with peverify

// ignore some error codes that appear to be false positives
```
peverify <renamed>.dll /ignore=0x80131252,0x80131B2C /hresult /transparent
```